Cambios generales en la interfaz gráfica.
Cuando juega, se muestra el sorteo que se está jugando. Cuando edita o agrega un sorteo de chances, todo lo relacionado a premios adicionales se oculta. También cuando edita, se muestra el sorteo que está editando

+++ Proyecto/FormResultados.cs
@@ -51,7 +51,7 @@ namespace Proyecto
             int numeroFraccion = (int)nudNumeroFraccion.Value;
             int serieFraccion = (int)nudSerieFraccion.Value;
             int cantidadFracciones = (int)nudCantidadFracciones.Value;
-            float montoGanado = 0;
+            double montoGanado = 0;
             Sorteo sorteo = ObtenerSorteoSeleccionado(tipo, numeroSorteo);
             Boolean gano = false;
             if (sorteo != null)
@@ -66,7 +66,7 @@ namespace Proyecto
                         {
                             if(cantidadFracciones <= sorteo.cantidadFracciones)
                             {
-                                montoGanado = resultado.montoGanado * ((float)cantidadFracciones / sorteo.cantidadFracciones);
+                                montoGanado = resultado.montoGanado * ((double)cantidadFracciones / sorteo.cantidadFracciones);
                             }
                             else
                             {
